% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{add_target_field}
\alias{add_target_field}
\title{TODO}
\usage{
add_target_field(
  stands,
  proj_unit,
  proj_target,
  proj_target_multiplier,
  proj_id,
  land_base
)
}
\arguments{
\item{stands}{TODO}

\item{proj_unit}{TODO}

\item{proj_target}{TODO}

\item{proj_target_multiplier}{TODO}

\item{proj_id}{TODO}

\item{land_base}{TODO}
}
\value{
TODO
}
\description{
TODO
}
