% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{apply_treatment}
\alias{apply_treatment}
\title{TODO}
\usage{
apply_treatment(
  stands,
  treatment_type,
  treatment_threshold,
  stand_field,
  proj_id,
  proj_fixed_target,
  proj_fixed_area_target = NULL,
  proj_unit = NULL,
  proj_target = NULL,
  proj_target_multiplier = NULL
)
}
\arguments{
\item{stands}{TODO}

\item{treatment_type}{TODO}

\item{treatment_threshold}{TODO}

\item{stand_field}{TODO}

\item{proj_id}{TODO}

\item{proj_fixed_target}{TODO}

\item{proj_fixed_area_target}{TODO}

\item{proj_unit}{TODO}

\item{proj_target}{TODO}

\item{proj_target_multiplier}{TODO}
}
\value{
TODO
}
\description{
TODO
}
