% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{attainment_chart_by_target_treated}
\alias{attainment_chart_by_target_treated}
\title{TODO}
\usage{
attainment_chart_by_target_treated(
  results_data,
  pcp_field,
  target_field,
  priority
)
}
\arguments{
\item{results_data}{TODO}

\item{pcp_field}{TODO}

\item{target_field}{TODO}

\item{priority}{TODO}
}
\value{
TODO
}
\description{
TODO
}
