% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{attainment_data_format}
\alias{attainment_data_format}
\title{TODO}
\usage{
attainment_data_format(attainment_data)
}
\arguments{
\item{attainment_data}{The prepared data from from attainment_data_filter}
}
\value{
TODO
}
\description{
TODO
}
