% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{build_dynamic_projects}
\alias{build_dynamic_projects}
\title{ForSysR wrapper for running patchmax}
\usage{
build_dynamic_projects(
  stands,
  proj_target_field,
  stand_id_field,
  proj_id_field,
  stand_area_field,
  stand_threshold,
  patchmax_proj_size,
  patchmax_proj_size_min,
  patchmax_proj_number,
  proj_target_value,
  proj_target_min_value,
  global_ceiling_field,
  global_ceiling_value,
  patchmax_SDW,
  patchmax_EPW,
  patchmax_exclusion_limit,
  patchmax_sample_frac,
  patchmax_sample_seed,
  patchmax_verbose,
  patchmax_adj_method = "queen",
  patchmax_adj_network = NULL
)
}
\value{
list
}
\description{
ForSysR wrapper for running patchmax
}
