% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions_external.R
\name{calculate_pcp}
\alias{calculate_pcp}
\title{Calculate pcp for specified fields}
\usage{
calculate_pcp(stands, fields = NULL)
}
\arguments{
\item{stands}{stand data}

\item{fields}{vector of character field names to calculate spm values}
}
\description{
Calculate pcp for specified fields
}
