% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions_aux.R
\name{create_output_directory}
\alias{create_output_directory}
\title{Create directory saving forsys output data}
\usage{
create_output_directory(relative_output_path, run_with_shiny, overwrite_output)
}
\arguments{
\item{run_with_shiny}{???}

\item{overwrite_output}{logical. Whether to first delete existing data in directory}

\item{relative}{relative path where new directory is placed}
}
\description{
Create directory saving forsys output data
}
