% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{cumulative_attainment_chart}
\alias{cumulative_attainment_chart}
\title{TODO}
\usage{
cumulative_attainment_chart(
  results_data,
  priority,
  constraint_field,
  secondary_effects = c()
)
}
\arguments{
\item{results_data}{The results data frame}

\item{priority}{The NAME of the main priority. This will be the filter for weight, and will add ETrt_*_PCP to access the value field}

\item{constraint_field}{Field used to constrain simulation. Usually some form of area.}

\item{secondary_effects}{(optional) Any other NAMES to select. It will automatically add ETrt_*_PCP to the names}
}
\value{
TODO
}
\description{
TODO
}
