% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_auxiliary_functions.R
\name{edit_json_config}
\alias{edit_json_config}
\title{Edit json config file in RStudio}
\usage{
edit_json_config(json)
}
\arguments{
\item{json}{}
}
\description{
Edit json config file in RStudio
}
