% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{filter_stands}
\alias{filter_stands}
\title{Filter data}
\usage{
filter_stands(stands, filter_txt)
}
\arguments{
\item{stands}{Data table to filter}

\item{filter_txt}{Boolean statement as character string}
}
\description{
Filter data
}
