% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_run.R
\docType{package}
\name{forsys}
\alias{forsys}
\title{forsys: A package for objective optimization of forest planning}
\description{
ForSys is a multi-objective spatial prioritization system that was designed to explore how
specific investment strategies aimed at improving forest conditions change under a range of
management constraints to achieve specific outcomes and understand tradeoffs. Forsys.app
modernizes how the USFS and our partners identify, prioritize, and implement treatment
activities across the landscape by using cloud-based, geospatial technology to explore
ForSys is a multi-objective spatial prioritization system that was designed to explore how
specific investment strategies aimed at improving forest conditions change under a range of
management constraints to achieve specific outcomes and understand tradeoffs. Forsys.app
modernizes how the USFS and our partners identify, prioritize, and implement treatment
activities across the landscape by using cloud-based, geospatial technology to explore
and rapidly visualize various management scenarios and treatment optimization decisions.
}
\section{forsys functions}{

run
}

