% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{get_result_priorities}
\alias{get_result_priorities}
\title{Get a list of priorities used in the run. This takes a results file (loaded object)
and parses it, so a loaded scenario isn't necessary. It looks for the ETrt*PCP columns}
\usage{
get_result_priorities(results_data)
}
\arguments{
\item{results_data}{TODO}
}
\value{
List of priorities identified in results file
}
\description{
Get a list of priorities used in the run. This takes a results file (loaded object)
and parses it, so a loaded scenario isn't necessary. It looks for the ETrt*PCP columns
}
