% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{get_result_targets}
\alias{get_result_targets}
\title{Right now this is a dumb function that returns all ETrt columns. Ideally we should have
logic that detects only the targets in the output file}
\usage{
get_result_targets(results_data)
}
\arguments{
\item{results_data}{TODO}
}
\value{
List of priorities identified in results file
}
\description{
Right now this is a dumb function that returns all ETrt columns. Ideally we should have
logic that detects only the targets in the output file
}
