% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{get_results_columns}
\alias{get_results_columns}
\title{Pass-through function while we shore up the graphing logic. This just sends ALL column
names back to the caller}
\usage{
get_results_columns(results_data)
}
\arguments{
\item{results_data}{TODO}
}
\value{
List of priorities identified in results file
}
\description{
Pass-through function while we shore up the graphing logic. This just sends ALL column
names back to the caller
}
