% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{list_attainment_targets}
\alias{list_attainment_targets}
\title{Get a list of 'y' columns, calculated in attainment_data_filter, which show
cumulative attainment in terms of PCP}
\usage{
list_attainment_targets(results_data)
}
\arguments{
\item{results_data}{The results data frame}
}
\value{
List of attainment columns
}
\description{
Get a list of 'y' columns, calculated in attainment_data_filter, which show
cumulative attainment in terms of PCP
}
