% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_scenario_functions.R
\name{list_scenarios}
\alias{list_scenarios}
\title{List json files in the configs folder, if it exists.}
\usage{
list_scenarios()
}
\value{
A list of found config files
}
\description{
List json files in the configs folder, if it exists.
}
