% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_run.R
\name{load_data}
\alias{load_data}
\title{Run the ForSys treatment planner. Either provide parameters, or define parameters
in a config file and pass the name of the file to this run function.}
\usage{
load_data(data_path)
}
\arguments{
\item{data_path}{Relative path to a config file that defines needed parameters}
}
\value{
A datatable with the weighted values for the priorities in the \code{priorityList}.
}
\description{
Run the ForSys treatment planner. Either provide parameters, or define parameters
in a config file and pass the name of the file to this run function.
}
