% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{load_dataset}
\alias{load_dataset}
\title{Load the input dataset. Supports both CSV and DBF.}
\usage{
load_dataset(path_to_file)
}
\arguments{
\item{path_to_file}{Path to an input dataset}
}
\value{
Loaded data.table from the input dataset
}
\description{
Load the input dataset. Supports both CSV and DBF.
}
