% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R,
%   R/forsys_scenario_functions.R
\name{load_json_config}
\alias{load_json_config}
\title{Load stored in JSON config into the current enviroment}
\usage{
load_json_config(json_filename)

load_json_config(json_filename)
}
\arguments{
\item{json_filename}{character string of forsys json config}
}
\description{
Load stored in JSON config into the current enviroment

Load stored in JSON config into the current enviroment
}
