% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{load_R_config}
\alias{load_R_config}
\title{TODO need a description for this function}
\usage{
load_R_config(config_file)
}
\arguments{
\item{config_file}{TODO}
}
\description{
TODO need a description for this function
}
