% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{make_thresholds}
\alias{make_thresholds}
\title{Threshold string statement parser}
\usage{
make_thresholds(txt)
}
\arguments{
\item{txt}{Vector of Boolean string statements to parse}
}
\description{
Threshold string statement parser
}
