% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{parse_priority_name}
\alias{parse_priority_name}
\title{Get a list of priorities used in the run. This takes a results file (loaded object)
and parses it, so a loaded scenario isn't necessary. It looks for the ETrt*PCP columns}
\usage{
parse_priority_name(column_name)
}
\arguments{
\item{column_name}{The column to parse out the priority name from}
}
\value{
Parsed priority name which can be used to match a Pr_ column
}
\description{
Get a list of priorities used in the run. This takes a results file (loaded object)
and parses it, so a loaded scenario isn't necessary. It looks for the ETrt*PCP columns
}
