% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{parse_thresholds}
\alias{parse_thresholds}
\title{TODO}
\usage{
parse_thresholds(field, operator, value)
}
\arguments{
\item{field}{TODO}

\item{operator}{TODO}

\item{value}{TODO}
}
\value{
TODO
}
\description{
TODO
}
