% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_auxiliary_functions.R
\name{print_json_config}
\alias{print_json_config}
\title{Print json config file to console}
\usage{
print_json_config(json)
}
\arguments{
\item{json}{}
}
\description{
Print json config file to console
}
