% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{printSpecsDocument}
\alias{printSpecsDocument}
\title{TODO add documentation for the specs function}
\usage{
printSpecsDocument(subunit, priorities, timber_threshold, volume_constraint)
}
\description{
TODO add documentation for the specs function
}
