% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{priority_column_name}
\alias{priority_column_name}
\title{From a priority (i.e. TVMBF) and get the associated Pr_ column}
\usage{
priority_column_name(results_data, priority)
}
\arguments{
\item{results_data}{The results data frame}

\item{priority}{TODO}
}
\value{
List of priorities identified in results file
}
\description{
From a priority (i.e. TVMBF) and get the associated Pr_ column
}
