% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{priority_etrt_pcp_name}
\alias{priority_etrt_pcp_name}
\title{Take a priority string and append ETrt_*_PCP to it}
\usage{
priority_etrt_pcp_name(priority)
}
\arguments{
\item{priority}{The name of a priority to turn into a ETrt_*_PCP string}
}
\value{
The ETrt_priority_PCP string
}
\description{
Take a priority string and append ETrt_*_PCP to it
}
