% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{production_frontiers_chart}
\alias{production_frontiers_chart}
\title{TODO}
\usage{
production_frontiers_chart(
  results_data,
  proj_field,
  x_field,
  y_field,
  constraint_field
)
}
\arguments{
\item{results_data}{TODO}

\item{proj_field}{TODO}

\item{x_field}{TODO}

\item{y_field}{TODO}

\item{constraint_field}{TODO}
}
\value{
TODO
}
\description{
TODO
}
