% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_scenario_functions.R
\name{read_save_file}
\alias{read_save_file}
\title{Load a json config file written from the write_save_file function.}
\usage{
read_save_file(filename = "")
}
\arguments{
\item{filename}{Relative path and filename to the scenario json}
}
\value{
A vector of input choices
}
\description{
Load a json config file written from the write_save_file function.
}
