% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{results_columns}
\alias{results_columns}
\title{Pass-through function while we shore up the graphing logic. This just sends
ALL column names back to the caller}
\usage{
results_columns(results_data)
}
\arguments{
\item{results_data}{The results data frame}
}
\value{
List of priorities identified in results file
}
\description{
Pass-through function while we shore up the graphing logic. This just sends
ALL column names back to the caller
}
