% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{Custom scale for the boxplot, this implements an inverted logrithmic scale}
\usage{
reverselog_trans(base = exp(1))
}
\arguments{
\item{base}{The logrithmic base to use}
}
\description{
Custom scale for the boxplot, this implements an inverted logrithmic scale
}
