% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{select_simple_greedy_algorithm}
\alias{select_simple_greedy_algorithm}
\title{Select subunits to treat based on a given priority.}
\usage{
select_simple_greedy_algorithm(
  dt = dbfFile,
  grouped_by = "PA_ID",
  prioritize_by = "TVMBF_SPM",
  tally_by = "AREA_HA",
  grouped_target = "AREA_PA10P"
)
}
\arguments{
\item{dt}{A data table with all the subunits and attributes.}

\item{grouped_by}{The management objective that is being prioritized.}

\item{prioritize_by}{The stand value for the priority that is used to constrain the results.
Typically Area or Volume.}

\item{tally_by}{The variable that is summed to meet the \code{tally_target}.}

\item{group_target}{The variable that contains the subunit values which are summed to meet the constraint value.}

\item{filter_by}{The variable that contains the values that are filtered by the thresholdVal.
i.e. this value must be greater than the threshold value.}

\item{filter_threshold}{The}
}
\value{
The selected stands from \code{df}, ordered by \code{prioritize_by}, and selected until the sum of \code{tally_by} is as close to
\code{group_target} as possible.
}
\description{
Select subunits to treat based on a given priority.
}
