% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{select_stands}
\alias{select_stands}
\title{Select subunits to treat based on a given priority.}
\usage{
select_stands(dt, grouped_by, prioritize_by, constrain_by, constraint_limit)
}
\arguments{
\item{dt}{data with all the subunits and attributes.}

\item{grouped_by}{field name with project or planning area ids}

\item{prioritize_by}{field name of priority to be maximized}

\item{constrain_by}{field name of constraint (e.g., are)}

\item{constraint_limit}{value of constraint not be exceeded}
}
\value{
The selected stands from \code{df}, ordered by \code{prioritize_by},
and selected until the sum of \code{tally_by} is as close to
\code{group_target} as possible.
}
\description{
Select subunits to treat based on a given priority.
}
