% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{set_fixed_area_target}
\alias{set_fixed_area_target}
\title{TODO}
\usage{
set_fixed_area_target(stands, fixed_area_target)
}
\arguments{
\item{stands}{TODO}

\item{fixed_area_target}{TODO}
}
\value{
TODO
}
\description{
TODO
}
