% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{set_fixed_target}
\alias{set_fixed_target}
\title{TODO}
\usage{
set_fixed_target(stands, target_value)
}
\arguments{
\item{stands}{TODO}

\item{target_value}{TODO}
}
\value{
TODO
}
\description{
TODO
}
