% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{set_percentage_area_target}
\alias{set_percentage_area_target}
\title{TODO}
\usage{
set_percentage_area_target(stands, proj_target, proj_target_multiplier)
}
\arguments{
\item{stands}{TODO}

\item{proj_target}{TODO}

\item{proj_target_multiplier}{TODO}
}
\value{
TODO
}
\description{
TODO
}
