% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{set_treatment_target}
\alias{set_treatment_target}
\title{Threshold string statement parser}
\usage{
set_treatment_target(
  stands,
  proj_id_field,
  proj_fixed_target,
  proj_target_field = NULL,
  proj_target_value = NULL
)
}
\arguments{
\item{stands}{TODO}

\item{proj_id_field}{TODO}

\item{proj_fixed_target}{TODO}

\item{proj_target_field}{TODO}

\item{proj_target_value}{TODO}
}
\description{
Threshold string statement parser
}
