% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{set_up_priorities}
\alias{set_up_priorities}
\title{Title}
\usage{
set_up_priorities(stands, w, priorities, weights)
}
\arguments{
\item{stands}{}

\item{w}{}

\item{priorities}{}

\item{weights}{}
}
\description{
Title
}
