% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{set_up_priorities_helper}
\alias{set_up_priorities_helper}
\title{TODO}
\usage{
set_up_priorities_helper(stands, i, weight, priority)
}
\arguments{
\item{stands}{TODO}

\item{i}{TODO}

\item{weight}{TODO}

\item{priority}{TODO}
}
\value{
TODO
}
\description{
TODO
}
