% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{set_up_treatment_types}
\alias{set_up_treatment_types}
\title{Hack the area target - can be set in the shapefile.
This code may be updated for looping multiple treatment types.
It should do the same thing that Pedro is working on within a single run
instead of wrapped.}
\usage{
set_up_treatment_types(stands, args = NULL)
}
\arguments{
\item{stands}{TODO}

\item{args}{TODO}
}
\value{
TODO
}
\description{
Hack the area target - can be set in the shapefile.
This code may be updated for looping multiple treatment types.
It should do the same thing that Pedro is working on within a single run
instead of wrapped.
}
