% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{set_variable_target}
\alias{set_variable_target}
\title{TODO}
\usage{
set_variable_target(stands, group_by, target_field, multiplier)
}
\arguments{
\item{stands}{TODO}

\item{group_by}{TODO}

\item{target_field}{TODO}

\item{multiplier}{TODO}
}
\value{
TODO
}
\description{
TODO
}
