% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{stacked_barchart}
\alias{stacked_barchart}
\title{TODO}
\usage{
stacked_barchart(
  subset_data,
  priority,
  proj_field,
  constraint_field,
  group_field
)
}
\arguments{
\item{subset_data}{The subset data results data frame}

\item{priority}{The NAME of the main priority. This will be the filter for
weight, and will add ETrt_*_PCP to access the value field}

\item{proj_field}{TODO}

\item{constraint_field}{Field used to constrain simulation. Usually some
form of area.}

\item{group_field}{TODO}
}
\value{
TODO
}
\description{
TODO
}
