% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{stand_filter}
\alias{stand_filter}
\title{Produce a set of stands that can be treated under a given criteria.}
\usage{
stand_filter(dt, filters)
}
\arguments{
\item{dt}{A data table with all stand information necessary to determine availability for a specific treatment type.}

\item{filters}{A list of strings that are used to filter the stands for treatment availability.}
}
\value{
The final data table with stands available for treatment.
}
\description{
Produce a set of stands that can be treated under a given criteria.
}
