% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{strip_str}
\alias{strip_str}
\title{For safety and sanity, take a string and rip off common output prefixes and
suffixes}
\usage{
strip_str(str_name)
}
\arguments{
\item{str_name}{The string to strip}
}
\value{
The stripped string
}
\description{
For safety and sanity, take a string and rip off common output prefixes and
suffixes
}
