% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{summarize_projects}
\alias{summarize_projects}
\title{Summarize project data}
\usage{
summarize_projects(
  selected_stands,
  stands_data,
  stand_id_field,
  stand_area_field,
  proj_id_field,
  scenario_output_grouping_fields,
  scenario_output_fields
)
}
\arguments{
\item{selected_stands}{}

\item{stands_data}{}

\item{stand_id_field}{}

\item{stand_area_field}{character. Name of variable containnig the stand area}

\item{proj_id_field}{}

\item{scenario_output_grouping_fields}{}

\item{scenario_output_fields}{}
}
\value{
list containing summary for selected projects and stands
}
\description{
Summarize project data
}
