% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{theme_toggle}
\alias{theme_toggle}
\title{Toggle function for switching between light and dark modes}
\usage{
theme_toggle(trigger)
}
\arguments{
\item{trigger}{TODO}
}
\value{
TODO
}
\description{
Toggle function for switching between light and dark modes
}
