% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{update_target}
\alias{update_target}
\title{Update target area for treatment after each activity}
\usage{
update_target(treated_stands, subunit, unit_area)
}
\arguments{
\item{treated_stands}{TODO}

\item{subunit}{TODO}

\item{unit_area}{TODO}
}
\value{
A table with the updated subunit targets for all planning areas that had treatments
}
\description{
Update target area for treatment after each activity
}
