% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{weight_priorities}
\alias{weight_priorities}
\title{Weight priorities for selection}
\usage{
weight_priorities(numPriorities, weights = c("1 1 1"))
}
\arguments{
\item{numPriorities}{TODO}

\item{weights}{TODO}
}
\value{
A datatable with the weighted values for the priorities in the \code{priorityList}.
}
\description{
Weight priorities for selection
}
