% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{weight_string_to_values}
\alias{weight_string_to_values}
\title{Inverse of weight_values_to_string}
\usage{
weight_string_to_values(weight_str)
}
\arguments{
\item{weight_str}{A string of minimum, maximum, and step weights}
}
\value{
A list of values cast as integers
}
\description{
Inverse of weight_values_to_string
}
