% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_auxiliary_functions.R
\name{weight_values_to_string}
\alias{weight_values_to_string}
\title{weight_values_to_string}
\usage{
weight_values_to_string(min = 0, max = 5, step = 1)
}
\arguments{
\item{min}{Minimum value}

\item{max}{Maximum value}

\item{step}{Increment value}
}
\value{
String of step values
}
\description{
Takes discrete minimum, maximum, and step values and turns them into
the string that forsys expects
}
