% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_scenario_functions.R
\name{write_save_file}
\alias{write_save_file}
\title{Write a json config file with the user's selected parameters}
\usage{
write_save_file(...)
}
\description{
Write a json config file with the user's selected parameters
}
