% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{add_target_field}
\alias{add_target_field}
\title{Sum of objective metric by project}
\usage{
add_target_field(
  stands,
  proj_unit,
  proj_target,
  proj_target_multiplier,
  stand_group_by
)
}
\arguments{
\item{stands}{Data table to filter}

\item{stand_group_by}{Field name to group by}
}
\description{
Sum of objective metric by project
}
