% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{apply_treatment}
\alias{apply_treatment}
\title{Select stands for treatment based on project stand thresholds and
threshold/targets}
\usage{
apply_treatment(
  stands,
  stand_id_field,
  proj_id_field,
  proj_objective = "weightedPriority",
  proj_target_field,
  proj_target = "master_target",
  treatment_name = NULL
)
}
\arguments{
\item{stands}{TODO}

\item{stand_id_field}{TODO}

\item{proj_id_field}{TODO}

\item{proj_objective}{TODO}

\item{proj_target_field}{TODO}

\item{proj_target}{TODO}

\item{treatment_name}{TODO}
}
\value{
TODO
}
\description{
Select stands for treatment based on project stand thresholds and
threshold/targets
}
