% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{attainment_chart_by_target_treated}
\alias{attainment_chart_by_target_treated}
\title{TODO}
\usage{
attainment_chart_by_target_treated(
  results_data,
  pcp_fields,
  constraint_field,
  priority
)
}
\arguments{
\item{results_data}{The results data frame, probably starts with 'proj_'}

\item{pcp_fields}{PCPs to chart, will be the Y axis values.}

\item{constraint_field}{Field used to constrain simulation. Usually some form of area.}

\item{priority}{The weight of priority to select. Different combinations of weights result in different outcomes}
}
\value{
TODO
}
\description{
TODO
}
