% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{attainment_data_format}
\alias{attainment_data_format}
\title{TODO}
\usage{
attainment_data_format(results_data)
}
\arguments{
\item{results_data}{The results data frame}
}
\value{
TODO
}
\description{
TODO
}
