% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{build_static_projects}
\alias{build_static_projects}
\title{Wrapper for selecting static stands}
\usage{
build_static_projects(stands, args)
}
\arguments{
\item{args}{}
}
\description{
Wrapper for selecting static stands
}
