% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{calculate_spm_pcp}
\alias{calculate_spm_pcp}
\title{TODO}
\usage{
calculate_spm_pcp(stands, filter, fields)
}
\arguments{
\item{stands}{TODO}

\item{filter}{TODO}

\item{fields}{TODO}
}
\value{
TODO
}
\description{
TODO
}
