% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{calculate_spm_pcp}
\alias{calculate_spm_pcp}
\title{Add spm and pcp values for specified fields}
\usage{
calculate_spm_pcp(stands, fields = NULL)
}
\arguments{
\item{stands}{data.table of stands}

\item{fields}{vector of character field names to calculate pcm & spm values}
}
\description{
Add spm and pcp values for specified fields
}
