% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{compile_planning_areas_and_stands}
\alias{compile_planning_areas_and_stands}
\title{TODO
TODO this should go in _results}
\usage{
compile_planning_areas_and_stands(
  stands,
  unique_weights,
  group_by,
  output_fields
)
}
\arguments{
\item{stands}{TODO}

\item{unique_weights}{TODO}

\item{group_by}{TODO}

\item{output_fields}{TODO}
}
\value{
TODO
}
\description{
TODO
TODO this should go in _results
}
