% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_external_functions.R
\name{filter_stands}
\alias{filter_stands}
\title{Filter data}
\usage{
filter_stands(stands, filter_txt = NULL, verbose = TRUE)
}
\arguments{
\item{stands}{Data table to filter}

\item{filter_txt}{Boolean statement as character string}

\item{verbose}{Boolean statement to report filtered results}
}
\description{
Used to filter stands to a specific criteria, often because certain stands
are considered outside the study area or otherwise excluded (e.g., wilderness
area, private lands, etc.)
}
