% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{project_boxplot}
\alias{project_boxplot}
\title{TODO}
\usage{
project_boxplot(results_data, proj_field, priority, constraint_field)
}
\arguments{
\item{results_data}{The results data frame}

\item{proj_field}{The planning area or project ID}

\item{priority}{The NAME of the main priority. This will be the filter for
weight, and will add ETrt_*_PCP to access the value field}

\item{constraint_field}{Field used to constrain simulation. Usually some
form of area.}
}
\value{
TODO
}
\description{
TODO
}
