% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForSys.R
\name{run}
\alias{run}
\title{Run the ForSys treatment planner. Either provide parameters, or define parameters
in a config file and pass the name of the file to this run function.}
\usage{
run(
  config_file = "",
  scenario_name = "",
  input_standfile = "",
  write_stand_outputs = FALSE,
  stand_field = "Cell_ID",
  pcp_spm = c(),
  land_base = "",
  priorities = c(),
  stand_group_by = "",
  pa_target = "",
  pa_unit = "",
  pa_target_multiplier = 0.15,
  nesting = FALSE,
  nesting_group_by = NULL,
  nesting_target = NULL,
  nesting_unit = NULL,
  nesting_target_multiplier = 1,
  weighting_values = "0 5 1",
  thresholds = c("Manageable man_alldis == 1"),
  include_stands = c("man_alldis == 1"),
  output_fields = c("AREA_HA", "TVMBF_STND", "TVMBF_PCP", "HUSUM_STND", "HUSUM_PCP"),
  grouping_variables = c("PA_ID", "Owner"),
  fixed_target = FALSE,
  fixed_area_target = 2000,
  overwrite_output = TRUE,
  spatial_optimization = FALSE,
  shiny_output = FALSE
)
}
\arguments{
\item{config_file}{Relative path to a config file that defines needed parameters}

\item{scenario_name}{A name for this scenario}

\item{input_standfile}{TODO}

\item{write_stand_outputs}{TODO}

\item{stand_field}{TODO}

\item{pcp_spm}{PCP and SPM values will be calculated for these variables. This should include the priorities and any value outputs.}

\item{land_base}{The land base is the area that is used to calculate the PCP and SPM values.
It is currently a single, binary variable that must be computed prior to running the ForSysR script.
A blank field means all lands are included in the calculation.}

\item{priorities}{Priorities are named here. If only one priority exists, only a weight of one will be used.}

\item{stand_group_by}{TODO}

\item{pa_target}{TODO}

\item{pa_unit}{TODO}

\item{pa_target_multiplier}{TODO}

\item{nesting}{TODO}

\item{nesting_group_by}{TODO}

\item{nesting_target}{TODO}

\item{nesting_target_multiplier}{TODO}

\item{weighting_values}{Defines the weights and integer steps between weights. The values are for min, max, and step.}

\item{thresholds}{Thresholds are defined by type (the first value in the string). The current code only uses one type (Commercial).}

\item{include_stands}{This defines global threshold values to include stands - i.e. for any threshold type.}

\item{output_fields}{This should include the desired fields for the planning area treatment files. Planning area id,
priority weights and treatment rank are added automatically.}

\item{grouping_variables}{Include the smaller and larger groups here for grouping of treated stands.}

\item{fixed_target}{Set to have either a fixed area target (TRUE) or a variable area target (FALSE)}

\item{fixed_area_target}{TODO}

\item{overwrite_output}{Toggle to overwrite existing output files}
}
\value{
A datatable with the weighted values for the priorities in the \code{priorityList}.
}
\description{
Run the ForSys treatment planner. Either provide parameters, or define parameters
in a config file and pass the name of the file to this run function.
}
