% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForSys.R
\name{run}
\alias{run}
\title{ForSysR}
\usage{
run(
  config_file = NULL,
  stand_data = NULL,
  stand_data_filename = "",
  stand_id_field = "",
  stand_pcp_spm = NULL,
  stand_area_field = NULL,
  global_threshold = NULL,
  stand_threshold = NULL,
  proj_id_field = "proj_id",
  proj_fixed_target = FALSE,
  proj_target_field = "",
  proj_target_value = NULL,
  proj_treatment_name = "",
  scenario_priorities = NULL,
  scenario_weighting_values = NULL,
  run_with_shiny = FALSE,
  run_with_patchmax = FALSE,
  run_with_fire = FALSE,
  fire_intersect_table = NULL,
  fire_intersect_table_filename = NULL,
  fire_planning_years = 1,
  fire_annual_target_field = NULL,
  fire_annual_target = NA,
  fire_dynamic_forsys = FALSE,
  fire_random_projects = FALSE,
  patchmax_stnd_adj = NULL,
  patchmax_stnd_adj_filename = NULL,
  patchmax_proj_number = 1,
  patchmax_proj_size = Inf,
  patchmax_proj_size_slack = NULL,
  patchmax_candidate_min_size = NULL,
  patchmax_sample_n = NULL,
  patchmax_sample_seed = NULL,
  scenario_output_fields = NULL,
  scenario_output_grouping_fields = NULL,
  scenario_name = "",
  scenario_write_tags = NULL,
  overwrite_output = TRUE
)
}
\arguments{
\item{config_file}{relative path to a config file that defines needed parameters}

\item{stand_data}{input stand dataset}

\item{stand_data_filename}{path to the input dataset (if not passed)}

\item{stand_id_field}{The field in the stand_data_filename which is a unique ID for each stand}

\item{stand_pcp_spm}{PCP and SPM values will be calculated for these variables. This should include the priorities and any value outputs.}

\item{stand_area_field}{string of field containing stand area}

\item{global_threshold}{Boolean statement passed as a string used to define stands within the scenario.
Excluded stands are not considered part of the problem so are not used to calculate PCP or ESum values.}

\item{stand_threshold}{Boolean statement passed as a string and used as threshold for whether stands are counted towards project objective}

\item{proj_id_field}{string of field in the stand_data_filename that indicates which project or planning area a stand belongs to.}

\item{proj_fixed_target}{logical describing if target is fixed or relative}

\item{proj_target_field}{string of stand field used as target constraint}

\item{proj_target_value}{numeric value for target constraint, either an fixed
value (if proj_fixed_target is TRUE) or a value between 0 and 1 (if proj_fixed_target is FALSE).}

\item{proj_treatment_name}{optional string appended to output used to name treatment}

\item{scenario_priorities}{Priorities are named here. If only one priority exists, only a weight of one will be used.}

\item{scenario_weighting_values}{string of 3 integers separated by a space that defines the weighting min, max, and step.}

\item{run_with_shiny}{logical whether run was called from within shiny.}

\item{run_with_patchmax}{logical whether PatchMax should be used for building projects}

\item{run_with_fire}{logical whether to forsys alongside fire}

\item{fire_intersect_table}{data frame listing stands affected by fire by year}

\item{fire_intersect_table_filename}{data frame listing stands affected by fire by year}

\item{fire_planning_years}{number of years to run forsys}

\item{fire_annual_target_field}{TODO}

\item{fire_annual_target}{TODO}

\item{fire_dynamic_forsys}{logical. Prevent burnt stands from being selected if TRUE}

\item{fire_random_projects}{logical. Randomly shuffle project prioritization if TRUE}

\item{patchmax_stnd_adj}{igraph object describe patch adjacency}

\item{patchmax_stnd_adj_filename}{igraph object describe patch adjacency}

\item{patchmax_proj_number}{TODO}

\item{patchmax_proj_size}{Integer. Number of patchmax stands to estimate}

\item{patchmax_proj_size_slack}{Numeric between 0 and 1 represent percent of slack allowed in project size constraint}

\item{patchmax_candidate_min_size}{TODO}

\item{patchmax_sample_n}{number of nodes used to build projects}

\item{patchmax_sample_seed}{set seed to reproduce random node selection above}

\item{scenario_output_fields}{vector of field names to write out}

\item{scenario_output_grouping_fields}{Include the smaller and larger groups here for grouping of treated stands.}

\item{scenario_name}{name for this scenario}

\item{scenario_write_tags}{optional string appended to output used to describe scenario}

\item{overwrite_output}{logical whether to overwrite any existing output of the same name}
}
\value{
Forsys results with weightedPriority, treatmentRank, and weights
}
\description{
Primary function for running the ForSys treatment planner. Either provide parameters, or define parameters
in a config file and pass the name of the file to this run function.
}
