% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_run.R
\name{run}
\alias{run}
\title{ForSysR}
\usage{
run(
  config_file = NULL,
  return_outputs = FALSE,
  write_outputs = TRUE,
  overwrite_output = TRUE,
  stand_data = NULL,
  stand_data_filename = NULL,
  stand_id_field = NULL,
  stand_area_field = NULL,
  stand_threshold = NULL,
  global_threshold = NULL,
  proj_id_field = NULL,
  proj_fixed_target = TRUE,
  proj_target_field = NULL,
  proj_target_value = NULL,
  proj_target_min_value = -Inf,
  planning_years = 1,
  annual_target_field = NULL,
  annual_target = NA,
  scenario_name = NULL,
  scenario_priorities = NULL,
  scenario_weighting_values = "1 1 1",
  scenario_output_fields = NULL,
  scenario_output_grouping_fields = NULL,
  scenario_write_tags = NULL,
  run_with_shiny = FALSE,
  run_with_patchmax = FALSE,
  run_with_fire = FALSE,
  fire_intersect_table = NULL,
  fire_intersect_table_filename = NULL,
  fire_dynamic_forsys = FALSE,
  fire_random_projects = FALSE,
  patchmax_proj_number = 1,
  patchmax_proj_size = Inf,
  patchmax_proj_size_min = -Inf,
  patchmax_sample_frac = 0.1,
  patchmax_st_seed = NULL,
  patchmax_SDW = 0.5,
  patchmax_EPW = 0.5
)
}
\arguments{
\item{config_file}{Config file that defines \code{forsys::run} parameters. <\emph{character}>}

\item{return_outputs}{Return project and stand directly. <\emph{logical}>}

\item{write_outputs}{Write project and stand data to file. <\emph{logical}>}

\item{overwrite_output}{Overwrites existing output of the same name. <\emph{logical}>}

\item{stand_data}{Stand dataset. <\emph{data.frame, data.table, sf}>}

\item{stand_data_filename}{Path to the input dataset saved as CSV file. <\emph{character}>}

\item{stand_id_field}{Field name containing unique ID for each stand. <\emph{character}>}

\item{stand_area_field}{Field name containing stand area. <\emph{character}>}

\item{stand_threshold}{Boolean statement defining stands counted towards project objective <\emph{character}>}

\item{global_threshold}{Boolean statement used to specify which stands are within the scenario. <\emph{character}>}

\item{proj_id_field}{Field name indicating which planning area a stand belongs to. <\emph{character}>}

\item{proj_fixed_target}{Whether target is fixed or relative. <\emph{logical}>}

\item{proj_target_field}{Field name used as target constraint. <\emph{character}>}

\item{proj_target_value}{Target constraint: fixed value if \code{proj_fixed_target} is TRUE or 0 and 1 if FALSE. <\emph{numeric}>}

\item{proj_target_min_value}{Minimum valid target constraint. Only used if \code{run_with_patchmax} is TRUE.}

\item{planning_years}{Number of years to run forsys. <\emph{integer}>}

\item{annual_target_field}{Field name to use for calculating annual target. <\emph{character}>}

\item{annual_target}{Value of annual cumulative target. <\emph{numeric}>}

\item{scenario_name}{Name for this scenario. <\emph{character}>}

\item{scenario_priorities}{Scenario priorities. <\emph{character vector}>}

\item{scenario_weighting_values}{String of 3 integers separated by spaces defining weighting min, max, and step. <\emph{character}>}

\item{scenario_output_fields}{Field names to write out. <\emph{character vector}>}

\item{scenario_output_grouping_fields}{Field names for grouping of treated stands. <\emph{character vector}>}

\item{scenario_write_tags}{String appended to output used to describe scenario. <\emph{optional character}>}

\item{run_with_shiny}{Whether run was called from within shiny. <\emph{logical}>}

\item{run_with_patchmax}{Whether PatchMax should be used for building projects. <\emph{logical}>}

\item{run_with_fire}{Whether to forsys alongside fire. <\emph{logical}>}

\item{fire_intersect_table}{Stands affected by fire by year. <\emph{data.frame}>}

\item{fire_intersect_table_filename}{File name of csv listing stands affected by fire by year. <\emph{character}>}

\item{fire_dynamic_forsys}{Whether burnt stands are prevented from being selected. <\emph{logical}>}

\item{fire_random_projects}{Whether project prioritization is randomly shuffled. <\emph{logical}>}

\item{patchmax_proj_number}{Number of projects to build with patchmax. <\emph{integer}>}

\item{patchmax_proj_size}{Target area for each patchmax project. <\emph{integer}>}

\item{patchmax_proj_size_min}{Minimum valid project size when using patchmax. <\emph{numeric}>}

\item{patchmax_sample_frac}{Percent of stands to search. <\emph{numeric 0-1}>}

\item{patchmax_st_seed}{Specific stand IDs to search. <\emph{numeric/character vector}>}

\item{patchmax_SDW}{Stand distance weight parameter. Default is 0.5. <\emph{numeric 0-1}>}

\item{patchmax_EPW}{Stand exclusion weight parameter. Default is 0.5. <\emph{numeric 0-1}>}
}
\value{
list with selected stands, project summary, project summary by subgroup
}
\description{
Primary function for running the ForSys treatment planner. Either provide
parameters, or define parameters in a config file and pass the name of the
file to this run function.
}
