% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{select_simple_greedy_algorithm}
\alias{select_simple_greedy_algorithm}
\title{Select subunits to treat based on a given priority.}
\usage{
select_simple_greedy_algorithm(dt, grouped_by, prioritize_by, constrain_by)
}
\arguments{
\item{dt}{A data table with all the subunits and attributes.}

\item{grouped_by}{The management objective that is being prioritized.}

\item{prioritize_by}{The stand value for the priority that is used to
constrain the results. Typically Area or Volume.}

\item{constrain_by}{This is a list of lists that includes three fields for
each constraint. The first item in each list, default 'apply', is the name
of a binary field in which 1 = apply the constraint and 0 = do not apply
the constraint to a given stand. The second item in each list is the name
of the field that is summed to reach the constraint. The third item in each
list is the name of the constraining field.}
}
\value{
The selected stands from \code{df}, ordered by \code{prioritize_by},
and selected until the sum of \code{tally_by} is as close to
\code{group_target} as possible.
}
\description{
Select subunits to treat based on a given priority.
}
