% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{set_treatment_target}
\alias{set_treatment_target}
\title{Threshold string statement parser}
\usage{
set_treatment_target(
  stands,
  proj_id_field,
  proj_fixed_target,
  proj_target_field = NULL,
  proj_target_value = NULL
)
}
\arguments{
\item{stands}{Dataframe containing stand data}

\item{proj_id_field}{Field name containing project id}

\item{proj_fixed_target}{Logical whether target is fixed or relative}

\item{proj_target_field}{Field name containing target value to be summed}

\item{proj_target_value}{Numeric or percent (0 - 1) depending on \code{proj_fixed_target}}
}
\description{
Threshold string statement parser
}
