% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R
\name{set_up_priorities}
\alias{set_up_priorities}
\title{TODO}
\usage{
set_up_priorities(stands, w, priorities, weights)
}
\arguments{
\item{stands}{TODO}

\item{w}{TODO}

\item{priorities}{TODO}

\item{weights}{TODO}
}
\description{
TODO
}
