% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_results_functions.R
\name{tradeoff_analysis_chart}
\alias{tradeoff_analysis_chart}
\title{TODO}
\usage{
tradeoff_analysis_chart(
  results_data,
  proj_field,
  x_field,
  y_field,
  proj_n = 10
)
}
\arguments{
\item{results_data}{TODO}

\item{proj_field}{The planning area or project ID}

\item{x_field}{The first priority to look at. Should be the priority name, e.g. "TVMIN",
which will be parsed to "ETrt_TVMIN_PCP"}

\item{y_field}{The second priotity to look at}
}
\value{
TODO
}
\description{
TODO
}
