% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_functions.R, R/forsys_run.R
\name{version}
\alias{version}
\title{Print the version information for Forsys.R}
\usage{
version()

version()
}
\value{
A string representing the version of Forsys.R

A string representing the version of Forsys.R
}
\description{
Print the version information for Forsys.R

Print the version information for Forsys.R
}
