% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_scenario_functions.R
\name{write_save_file_helper}
\alias{write_save_file_helper}
\title{Take the input object from a shiny server function all at once and then call the write_save_file function
This function helps keep the server code clean.}
\usage{
write_save_file_helper(input, data_path)
}
\arguments{
\item{input}{input object from a shiny server function}

\item{data_path}{System path to save the output file to}
}
\value{
A serialized vector of input choices in json format
}
\description{
Take the input object from a shiny server function all at once and then call the write_save_file function
This function helps keep the server code clean.
}
